# cmd/api/usersBooks.go
package main

import (
	"errors"
	"net/http"
	"time"

	"github.com/svenrisse/bookshelf/internal/models"
	"github.com/svenrisse/bookshelf/internal/validator"
)

// createUserBookHandler godoc
//
//	@Summary		add a userBook
//	@Description	this is a connection between users and books
//	@Tags			userBooks
//	@Accept			json
//	@Produce		json
//	@Param			book	body		models.UserBook	true	"Create UserBook"
//	@Success		201		{object}	models.UserBook
//	@Failure		400
//	@Failure		422
//	@Failure		500
//	@Router			/user/books [post]
func (app *application) createUsersBooksHandler(w http.ResponseWriter, r *http.Request) {
	var input struct {
		BookID     int64     `json:"bookID"`
		UserID     int64     `json:"userID"`
		Read       bool      `json:"read"`
		Rating     float32   `json:"rating,omitempty"`
		ReviewBody string    `json:"reviewBody,omitempty"`
		ReadAt     time.Time `json:"readAt,omitempty"`
		ReviewedAt time.Time `json:"reviewedAt,omitempty"`
	}

	err := app.readJSON(w, r, &input)
	if err != nil {
		app.badRequestResponse(w, r, err)
		return
	}

	userBook := &models.UserBook{
		BookID:     input.BookID,
		UserID:     input.UserID,
		Read:       input.Read,
		Rating:     input.Rating,
		ReviewBody: input.ReviewBody,
		ReadAt:     input.ReadAt,
		ReviewedAt: input.ReviewedAt,
	}

	v := validator.New()
	if models.ValidateUserBook(v, userBook); !v.Valid() {
		app.failedValidationResponse(w, r, v.Errors)
		return
	}

	err = app.models.UserBook.Insert(userBook)
	if err != nil {
		app.serverErrorResponse(w, r, err)
		return
	}

	err = app.writeJSON(w, http.StatusCreated, envelope{"userBook": userBook}, nil)
	if err != nil {
		app.serverErrorResponse(w, r, err)
	}
}

func (app *application) deleteUsersBooksHandler(w http.ResponseWriter, r *http.Request) {
	id, err := app.readIDParam(r)
	if err != nil {
		app.notFoundResponse(w, r)
		return
	}

	err = app.models.UserBook.Delete(id)
	if err != nil {
		if errors.Is(err, models.ErrRecordNotFound) {
			app.notFoundResponse(w, r)
			return
		}
		app.serverErrorResponse(w, r, err)
		return
	}

	err = app.writeJSON(w, http.StatusOK, envelope{"message": "UserBook successfully deleted"}, nil)
	if err != nil {
		app.serverErrorResponse(w, r, err)
	}
}

func (app *application) updateUsersBooksHandler(w http.ResponseWriter, r *http.Request) {
	id, err := app.readIDParam(r)
	if err != nil {
		app.notFoundResponse(w, r)
	}

	userBook, err := app.models.UserBook.Get(id)
	if err != nil {
		if errors.Is(err, models.ErrRecordNotFound) {
			app.notFoundResponse(w, r)
			return
		}
		app.serverErrorResponse(w, r, err)
		return
	}

	var input struct {
		Read       *bool
		Rating     *float32
		ReviewBody *string
		ReadAt     *time.Time
		ReviewedAt *time.Time
	}

	err = app.readJSON(w, r, &input)
	if err != nil {
		app.badRequestResponse(w, r, err)
		return
	}

	if input.Read != nil {
		userBook.Read = *input.Read
	}
	if input.Rating != nil {
		userBook.Rating = *input.Rating
	}
	if input.ReviewBody != nil {
		userBook.ReviewBody = *input.ReviewBody
	}
	if input.ReadAt != nil {
		userBook.ReadAt = *input.ReadAt
	}
	if input.ReviewedAt != nil {
		userBook.ReviewedAt = *input.ReviewedAt
	}

	v := validator.New()

	if models.ValidateUserBook(v, userBook); !v.Valid() {
		app.failedValidationResponse(w, r, v.Errors)
		return
	}

	err = app.models.UserBook.Update(userBook)
	if err != nil {
		if errors.Is(err, models.ErrEditConflict) {
			app.editConflictResponse(w, r)
			return
		}
		app.serverErrorResponse(w, r, err)
		return
	}

	err = app.writeJSON(w, http.StatusOK, envelope{"userBook": userBook}, nil)
	if err != nil {
		app.serverErrorResponse(w, r, err)
	}
}

func (app *application) listUsersBooksHandler(w http.ResponseWriter, r *http.Request) {
	var input struct {
		Title  string
		Genres []string
		Rating float32
		Read   bool
		models.Filters
	}

	v := validator.New()

	qs := r.URL.Query()

	input.Title = app.readString(qs, "title", "")
	input.Genres = app.readCSV(qs, "genres", []string{})
	input.Rating = app.readFloat(qs, "rating", 0, v)
	input.Read = app.readBool(qs, "read", false, v)
	input.Filters.Page = app.readInt(qs, "page", 1, v)
	input.Filters.PageSize = app.readInt(qs, "page_size", 20, v)
	input.Filters.Sort = app.readString(qs, "sort", "id")
	input.Filters.SortSafeList = []string{
		"title",
		"rating",
		"read_at",
		"reviewed_at",
		"published_at",
		"-title",
		"-rating",
		"-read_at",
		"-reviewed_at",
		"-published_at",
	}

	if models.ValidateFilters(v, input.Filters); v.Valid() {
		app.failedValidationResponse(w, r, v.Errors)
		return
	}

	userBooks, metadata, err := app.models.UserBook.List(
		input.Title,
		input.Genres,
		input.Rating,
		input.Read,
		input.Filters,
	)
	if err != nil {
		app.serverErrorResponse(w, r, err)
	}

	err = app.writeJSON(
		w,
		http.StatusOK,
		envelope{"userBooks": userBooks, "metadata": metadata},
		nil,
	)
	if err != nil {
		app.serverErrorResponse(w, r, err)
	}
}