# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is the API for the Bookshelf application.",
        "title": "Bookshelf API",
        "contact": {
            "name": "API Support",
            "email": "svenrisse0@gmail.com"
        },
        "version": "1.0"
    },
    "host": "bookshelf.svenrisse.com",
    "basePath": "/v1/",
    "paths": {
        "/user/books": {
            "post": {
                "description": "this is a connection between users and books",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "userBooks"
                ],
                "summary": "add a userBook",
                "parameters": [
                    {
                        "description": "Create UserBook",
                        "name": "book",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_svenrisse_bookshelf_internal_models.UserBook"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/github_com_svenrisse_bookshelf_internal_models.UserBook"
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "422": {
                        "description": "Unprocessable Entity"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/v1/auth/{provider}": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "users"
                ],
                "summary": "Let a User Login with given Provider",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Discord / Github / Google",
                        "name": "provider",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "302": {
                        "description": "Found"
                    }
                }
            }
        },
        "/v1/auth/{provider}/logout": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "users"
                ],
                "summary": "Let a User Logout with given Provider",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Discord / Github / Google",
                        "name": "provider",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "302": {
                        "description": "Found"
                    }
                }
            }
        },
        "/v1/books": {
            "post": {
                "description": "create book with fields",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "books"
                ],
                "summary": "Create a Book",
                "parameters": [
                    {
                        "description": "Add book",
                        "name": "book",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_svenrisse_bookshelf_internal_models.Book"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/github_com_svenrisse_bookshelf_internal_models.Book"
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "422": {
                        "description": "Unprocessable Entity"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/v1/books/{id}": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "books"
                ],
                "summary": "Get a Book",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Book ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_svenrisse_bookshelf_internal_models.Book"
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            },
            "patch": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "books"
                ],
                "summary": "Update a book by providing new values",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Book ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Provide Fields to change",
                        "name": "book",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_svenrisse_bookshelf_internal_models.Book"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_svenrisse_bookshelf_internal_models.Book"
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "409": {
                        "description": "Conflict"
                    },
                    "422": {
                        "description": "Unprocessable Entity"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        }
    },
    "definitions": {
        "github_com_svenrisse_bookshelf_internal_models.Book": {
            "description": "Book information",
            "type": "object",
            "properties": {
                "author": {
                    "type": "string",
                    "example": "J.R.R. Tolkien"
                },
                "genres": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    },
                    "example": [
                        "Fantasy",
                        "Epic",
                        "Children's literature"
                    ]
                },
                "pages": {
                    "type": "integer",
                    "example": 320
                },
                "title": {
                    "type": "string",
                    "example": "The Hobbit"
                },
                "year": {
                    "type": "integer",
                    "example": 1937
                }
            }
        },
        "github_com_svenrisse_bookshelf_internal_models.UserBook": {
            "description": "UserBook information",
            "type": "object",
            "properties": {
                "book_id": {
                    "type": "integer",
                    "example": 1
                },
                "rating": {
                    "type": "number",
                    "example": 4.5
                },
                "read": {
                    "type": "boolean",
                    "example": true
                },
                "read_at": {
                    "type": "string",
                    "example": "1975-08-19T23:15:30.000Z"
                },
                "review_body": {
                    "type": "string",
                    "example": "I loved this book!"
                },
                "reviewed_at": {
                    "type": "string",
                    "example": "1975-08-19T23:15:30.000Z"
                },
                "user_id": {
                    "type": "integer",
                    "example": 1
                }
            }
        }
    }
}